+++ src/main/java/fll/xml/ChallengeParser.java
@@ -282,12 +282,12 @@ public final class ChallengeParser {
           simpleGoals.put(name, element);
 
           // check initial values
-          final double initialValue = Utilities.FLOATING_POINT_NUMBER_FORMAT_INSTANCE.parse(element.getAttribute("initialValue"))
+          final double initialValue = Utilities.XML_FLOATING_POINT_NUMBER_FORMAT_INSTANCE.parse(element.getAttribute("initialValue"))
                                                                                      .doubleValue();
           if (ChallengeParser.isEnumeratedGoal(element)) {
             boolean foundMatch = false;
             for (final Element valueEle : new NodelistElementCollectionAdapter(element.getChildNodes())) {
-              final double score = Utilities.FLOATING_POINT_NUMBER_FORMAT_INSTANCE.parse(valueEle.getAttribute("score"))
+              final double score = Utilities.XML_FLOATING_POINT_NUMBER_FORMAT_INSTANCE.parse(valueEle.getAttribute("score"))
                                                                                   .doubleValue();
               if (FP.equals(score, initialValue, INITIAL_VALUE_TOLERANCE)) {
                 foundMatch = true;
@@ -299,9 +299,9 @@ public final class ChallengeParser {
             }
 
           } else {
-            final double min = Utilities.FLOATING_POINT_NUMBER_FORMAT_INSTANCE.parse(element.getAttribute("min"))
+            final double min = Utilities.XML_FLOATING_POINT_NUMBER_FORMAT_INSTANCE.parse(element.getAttribute("min"))
                                                                               .doubleValue();
-            final double max = Utilities.FLOATING_POINT_NUMBER_FORMAT_INSTANCE.parse(element.getAttribute("max"))
+            final double max = Utilities.XML_FLOATING_POINT_NUMBER_FORMAT_INSTANCE.parse(element.getAttribute("max"))
                                                                               .doubleValue();
             if (FP.lessThan(initialValue, min, INITIAL_VALUE_TOLERANCE)) {
               throw new InvalidInitialValue(String.format("Initial value for %s(%f) is less than min(%f)", name,