+++ src/main/java/fll/xml/PerformanceScoreCategory.java
@@ -192,7 +192,8 @@ public class PerformanceScoreCategory extends ScoreCategory {
 
     populateXml(doc, ele);
 
-    ele.setAttribute(MINIMUM_SCORE_ATTRIBUTE, Utilities.FLOATING_POINT_NUMBER_FORMAT_INSTANCE.format(mMinimumScore));
+    ele.setAttribute(MINIMUM_SCORE_ATTRIBUTE,
+                     Utilities.XML_FLOATING_POINT_NUMBER_FORMAT_INSTANCE.format(mMinimumScore));
 
     for (final Restriction restrict : mRestrictions) {
       final Element restrictEle = restrict.toXml(doc);